nes: cleanup: extract edit intent parsing
